+++ vue/src/components/pages/blogPohomovaKapustin/components/Album/AlbumTitle.vue
@@ -1,0 +1,18 @@
+<template>
+    <div class="album-title">
+        Альбомы 2023/24
+    </div>
+</template>
+    
+<script>
+</script>
+    
+<style scoped lang="less">
+.album-title {
+    color: white;
+    font-family: "Lucida Handwriting", cursive;
+    font-size: 30px;
+    letter-spacing: 5px;
+}
+</style>
+
