+++ vue/src/components/pages/blogPohomovaKapustin/components/Musician/MusicianCell.vue
@@ -1,0 +1,14 @@
+<template>
+    <div>
+      <img :src="musician.imageUrl" :alt="musician.name" class="album__image" />
+    </div>
+</template>
+    
+<script>
+
+</script>
+    
+<style scoped lang="less">
+
+</style>
+
